Moved the networking library to a better named socket library
