+++ sockets/SocketConsts.h
