+++ sockets/SocketManager.h
